refactor(select in googleinput): added native html select in GoogleInput instead of mui select
