+++ src/store/sagas/__test__/authSaga.spec.js
@@ -1,6 +1,7 @@
 import { recordSaga } from "../recordSaga";
 import { authenticateSaga } from "../authSaga";
 import { authenticate } from "../../actions/authorization";
+import {serverLogin} from "../../../api";
 
 
 jest.mock("../../../api", () => ({
@@ -10,6 +11,7 @@ jest.mock("../../../api", () => ({
 describe("authSaga", () => {
     describe("#AUTHENTICATE", () => {
       it("authenticates through api", async () => {
+        serverLogin.mockImplementation(async () => ({success: true, token: 123}))
         const dispatched = await recordSaga(
           authenticateSaga,
           authenticate("testlogin", "testpassword")
@@ -18,6 +20,9 @@ describe("authSaga", () => {
         expect(dispatched).toEqual([
           {
             type: "LOG_IN",
+            payload: {
+              token: 123
+            }
           },
         ]);
       });
